refactor(portal): share display shortcuts runtime helper
