+++ src/larvaworld/portal/runtime/__init__.py
@@ -1,0 +1,2 @@
+"""Shared runtime helpers for portal applications."""
+
